feat: add link command to allow linking after use
If you accidentally remove or write over the symlink at /usr/local/bin/ipfs `iim link` will restore it, pointing it at the current IPFS binary.

It's also useful in the case where a user does not have sufficient privileges to write to /usr/local/bin because it allows IPFS to be installed with a non-privileged account and then symlinked with `sudo iim link` afterwards.

License: MIT
Signed-off-by: Alan Shaw <alan.shaw@protocol.ai>